+++ lib/onlyoffice_documentserver_testing_framework.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative 'onlyoffice_documentserver_testing_framework/document_server_version'
 require_relative 'onlyoffice_documentserver_testing_framework/selenium_wrapper'
 require_relative 'onlyoffice_documentserver_testing_framework/test_instance_docs'
 require_relative 'onlyoffice_documentserver_testing_framework/version'